refactored test class to name with clear association to class it's testing.
